+++ lib/avx512_t1/kasumi_avx512.c
@@ -30,6 +30,5 @@
 #define AVX512
 #define CLEAR_SCRATCH_SIMD_REGS clear_scratch_xmms_avx
 #define KASUMI_F8_1_BUFFER      kasumi_f8_1_buffer_avx512
-#define KASUMI_F8_1_BUFFER_BIT  kasumi_f8_1_buffer_bit_avx512
 #define KASUMI_F9_1_BUFFER      kasumi_f9_1_buffer_avx512
 #include "include/kasumi_common_avx.h"