+++ lib/include/arch_avx512_type1.h
@@ -210,10 +210,6 @@ shake256_avx512(const uint8_t *input, const uint64_t inputByteLen, uint8_t *outp
 void
 kasumi_f8_1_buffer_avx512(const kasumi_key_sched_t *pCtx, const uint64_t IV, const void *pBufferIn,
                           void *pBufferOut, const uint32_t cipherLengthInBytes);
-void
-kasumi_f8_1_buffer_bit_avx512(const kasumi_key_sched_t *pCtx, const uint64_t IV,
-                              const void *pBufferIn, void *pBufferOut,
-                              const uint32_t cipherLengthInBits, const uint32_t offsetInBits);
 
 void
 kasumi_f9_1_buffer_avx512(const kasumi_key_sched_t *pCtx, const void *pBufferIn,