+++ lib/include/kasumi_interface.h
@@ -51,11 +51,6 @@ kasumi_f8_1_buffer_sse(const kasumi_key_sched_t *pCtx, const uint64_t IV, const 
                        void *pBufferOut, const uint32_t cipherLengthInBytes);
 
 void
-kasumi_f8_1_buffer_bit_sse(const kasumi_key_sched_t *pCtx, const uint64_t IV, const void *pBufferIn,
-                           void *pBufferOut, const uint32_t cipherLengthInBits,
-                           const uint32_t offsetInBits);
-
-void
 kasumi_f9_1_buffer_sse(const kasumi_key_sched_t *pCtx, const void *pBufferIn,
                        const uint32_t lengthInBytes, void *pDigest);